+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## Unreleased
 
+### Features
+
+- Add sentry-cli upload-dif action ([#89](https://github.com/getsentry/sentry-fastlane-plugin/pull/89))
+
+
 ## 1.8.3
 
 ### Fixes
 
 - Print sentry-cli level debug when verbose is set ([#88](https://github.com/getsentry/sentry-fastlane-plugin/pull/88))
 - Fix action to allow stripping prefix ([#87](https://github.com/getsentry/sentry-fastlane-plugin/pull/87))
+>>>>>>> master
 
 ## 1.8.2